+++ src/Cryptography/SharedSeparator.cs
@@ -1,0 +1,47 @@
+// ----------------------------------------------------------------------------
+// <copyright file="SharedSeparator.cs" company="Private">
+// Copyright (c) 2023 All Rights Reserved
+// </copyright>
+// <author>Sebastian Walther</author>
+// <date>05/27/2023 06:05:12 PM</date>
+// ----------------------------------------------------------------------------
+
+#region License
+// ----------------------------------------------------------------------------
+// Copyright 2019 Sebastian Walther
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+#endregion
+
+namespace SecretSharingDotNet.Cryptography
+{
+    internal static class SharedSeparator
+    {
+        /// <summary>
+        /// The separator between the X and Y coordinate
+        /// </summary>
+        internal const char CoordinateSeparator = '-';
+
+        /// <summary>
+        /// Separator array for <see cref="string.Split(char[])"/> method usage to avoid allocation of a new array.
+        /// </summary>
+        internal static readonly char[] CoordinateSeparatorArray = { CoordinateSeparator };
+    }
+
+}