+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.2] - 2023-09-16
+### Changed
+- FinitePoint: Use `ReadOnlySpan<char>` for `string` parameters in ctor and methods.
+
+### Deprecated
+- The Secret Legacy Mode is marked as deprecated and will be removed in one of the next releases.
+
+### Fixed
+- Fixed NETSDK1187 warning (see #151).  Locale 'de-de' is normalized to the standard locale 'de-DE'.
+
 ## [0.10.1] - 2023-05-08
 ### Fixed
 - Fixed BigIntCalculator's Equals method to avoid timing attacks. The slow equal implementation is used now.
@@ -182,6 +192,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `LICENSE.md`
 - Added `README.md`
 
+[0.10.2]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.8.0...v0.9.0